feat: temporary deviation from auth to create a basic UI
