+++ .gitignore
@@ -28,6 +28,7 @@ yarn-error.log*
 
 # local env files
 .env*.local
+__bak*
 
 # vercel
 .vercel
