+++ app/api/auth/[...nextauth]/authOptions.ts
@@ -1,0 +1,15 @@
+import { NextAuthOptions } from "next-auth";
+import SpotifyProvider from "next-auth/providers/spotify";
+
+const spotifyClientId = String(process?.env?.SPOTIFY_CLIENT_ID || "system-error-trigger-exception");
+const spotifyClientSecret = String(process?.env?.SPOTIFY_CLIENT_SECRET || "system-error-trigger-exception");
+
+
+export const authOptions: NextAuthOptions = {
+  providers: [
+    SpotifyProvider({
+      clientId: spotifyClientId,
+      clientSecret: spotifyClientSecret
+    })
+  ]
+}
