+++ app/layout.tsx
@@ -1,9 +1,5 @@
 import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import '../styles/globals.css'
-
-
-const inter = Inter({ subsets: ["latin"] });
+import "../styles/globals.css";
 
 export const metadata: Metadata = {
   title: { default: "Dashboard", template: "%s | My Website" },
@@ -17,17 +13,13 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <head>
-        <link
-          rel="preconnect"
-          href="https://fonts.gstatic.com"
-        />
-        <link
-          href="https://fonts.googleapis.com/css2?family=Montserrat:wght@300;400;500;600;700&display=swap"
-          rel="stylesheet"
-        />
-      </head>
-      <body className={inter.className}>{children}</body>
+      <head></head>
+      <body>
+        <div>
+          
+        </div>
+        {children}
+      </body>
     </html>
   );
 }
