--- app/types/next-auth.d.ts
@@ -1,8 +1,0 @@
-import NextAuth, { Profile as DefaultProfile } from "next-auth"
-
-declare module "next-auth" {
-  interface Profile {
-    country: string 
-    & DefaultProfile
-  }
-}
