--- components/CardItemGrid.tsx
@@ -1,3 +1,0 @@
-export default function CardItemGrid({ children }: any) {
-  return <div className="grid grid-cols-6 gap-6">{children}</div>;
-}
