+++ package.json
@@ -9,6 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@headlessui/react": "^1.7.18",
+    "@heroicons/react": "^2.1.1",
     "axios": "^1.6.5",
     "html-react-parser": "^1.4.4",
     "next": "^14.0.4",
